Calculo de combustible

+++ resources/views/servicio/servicios.blade.php
@@ -5,7 +5,7 @@
 <article class="col-md-12">
   <div class="panel panel-default">
     <div class="panel-heading">
-      Inicio Alerta
+      Ultimos servicios realizados
     </div>
     <div class="panel-body">
       <table  class="table table-bordered">
@@ -13,32 +13,29 @@
           <tr>
             <th class="text-center">Tipo de servicio</th>
             <th class="text-center">Direccion</th>
-            <th class="text-center">Hora de alarma</th>
+            <th class="text-center">Hora de regreso</th>
             <th colspan="2"></th>
           </tr>
         </thead>
         <tbody><!--Contenido de la tabla-->
           @foreach ($servicios as $servicio)
-            <tr>
-              <td class="text-center">{{$servicio->tipo_servicio_id}}</td>
-              <td class="text-center">{{$servicio->direccion}}</td>
-              <td class="text-center">{{$servicio->hora_alarma}}</td>
-              <td class="text-center">
-                {{ Form::open(['route' => ['servicio.destroy', $servicio->id], 'method' => 'delete']) }}
-                    <button type="submit" class="btn glyphicon glyphicon-trash eliminar"></button>
-                {{ Form::close() }}
-              </td>
-              <td class="text-center"><a class="glyphicon glyphicon-edit" href="{{ route('servicio.mostrar', $servicio->id) }}"></a></td>
-            </tr>
+            @if ($servicio->hora_regreso)
+              <tr>
+                <td class="text-center">{{$servicio->tipoServicio->nombre}}</td>
+                <td class="text-center">{{$servicio->direccion}}</td>
+                <td class="text-center">{{$servicio->hora_regreso}}</td>
+                <td class="text-center">
+                  {{ Form::open(['route' => ['servicio.destroy', $servicio->id], 'method' => 'delete']) }}
+                      <button type="submit" class="btn glyphicon glyphicon-trash eliminar"></button>
+                  {{ Form::close() }}
+                </td>
+                {{-- finalizarActivo implementar Update --}}
+                <td class="text-center"><a class="glyphicon glyphicon-edit" href="{{ route('servicio.finalizarActivo', $servicio->id) }}"></a></td>
+              </tr>
+            @endif
           @endforeach
         </tbody>
-          <tfoot>
-            <tr>
-              <td class="text-center" colspan="5"> Lista de bomberos activos </td>
-            </tr>
-          </tfoot>
-          <br>
-        </table>
+      </table>
     </div>
   </div>
 </article>
